\name{fanchart_percentiles}
\alias{fanchart_percentiles}
\title{Time series of percentiles}
\description{
Generates time series of percentiles of simulated scenarios.
}
\usage{
fanchart_percentiles(
  s,
  percentiles = c(0.1,0.25,0.5,0.75,0.9)
)
}
\arguments{
\item{s}{A \code{list} generated by function \code{\link{sim}}.}
\item{percentiles}{\code{numeric(n)}. Numeric vector with one or more percentiles, in decimal form. For example, for the 10\% and 90\% percentiles, use \code{c(0.1,0.9)}.}
}
\value{

A \code{list} containing object \code{the_percentiles}. The latter is also a \code{list}, but of \code{xts} objects. There will be one \code{xts} object for each variable in \code{s}. Such \code{xts} objects will have one time series (column) for each percentile in argument \code{percentiles}.

}
\examples{
## Step 1) generate yearly shocks using function "shocks_generator"
shocks <- shocks_generator(
  case = "ec",
  n_stochastic_scenarios = 100,
  realized_data = sample_series$realized_yearly_data,
  subintervals = "2009-12-01/2019-12-01",
  shocks_first_date = "2024-12-01",
  shocks_last_date = "2028-12-01",
  shocks_frequency = "year",
  i = "nominal interest rate",
  ti = "gdp deflator",
  g = "real gdp growth",
  p = "primary balance"
)

## Step 2) simulate the stochastic scenarios
s <- sim(
  first_date_to_simulate = "2024-12-01",
  last_date_to_simulate = "2028-12-01",
  realized_data = sample_series$realized_yearly_data,
  baseline_scenario = sample_series$baseline_yearly_scenario,
  baseline_as_median = TRUE,
  n_stochastic_scenarios = 100,
  shocks = shocks$shocks,
  d = "debt",
  i = "nominal interest rate",
  ti = "gdp deflator",
  r = "real interest rate (w.r.t. gdp deflator)",
  g = "real gdp growth",
  p = "primary balance"
)

## Step 3) computation of the 10%, 25%, 50%, 75% and 90% percentiles
fc <- fanchart_percentiles(
  s = s,
  percentiles = c(0.1,0.25,0.5,0.75,0.9)
)
}
